Sounds: added rain, changed step sounds and added water splashes. Made music silent until player hits spacebar. Other: moved enemies to outsidetilemaps so they despawn when inside.
I changed the footstep sounds so it will check if Rain-tiles are active or if the outside tiles are active and change the sounds accordingly. I had to bodge with enemy prefabs because you cannot reference scene objects.. So I tagged the objects I wanted to reference and then I searched them in code after initialization.

I made same kind of checks for rain sounds. In Rain tiles there's script that checks if outside tiles are active.

Also I upgraded the music sounds and propaganda sounds.